Fixes oversized starter kit hero
width set to 100vw calculates the wrong size in chrome/ff so I've removed the width property so that it's automatic. background position was also set to 50% on Y axis as most of the photos seem to be art-directed as centered. As it was, in a skinny window, you'd just see peoples' feet.